11 練習 Entity Framework 設計工具的「從資料庫更新模型」功能、12 套用模型類別的 MetadataType 類別，使用 EFGeneratePartialClass.tt

+++ MVC5Course/Controllers/ProductsController.cs
@@ -94,6 +94,10 @@ namespace MVC5Course.Controllers
         {
             var c = db.Product.Find(id);
 
+            if (c == null) {
+                return HttpNotFound();
+            }
+
             db.Product.Remove(c);
             db.SaveChanges();
